macosx compilation: ported ./m script + some intptr_t and some other casting 	modified:   src/m 	modified:   src/neoip_base/tools/dvar/subtype/dbl/neoip_dvar_dbl.hpp 	modified:   src/neoip_base/tools/dvar/subtype/int/neoip_dvar_int.hpp 	modified:   src/neoip_bt_ez/ezswarm/neoip_bt_ezswarm.cpp 	modified:   src/neoip_inet/base/address/ipport_addr/base/strlist_src/neoip_ipport_strlist_src.cpp 	modified:   src/neoip_ndiag/watch/neoip_ndiag_watch.cpp

+++ src/neoip_base/tools/dvar/subtype/dbl/neoip_dvar_dbl.hpp
@@ -35,7 +35,7 @@ public:
 	dvar_type_t	type()		const throw()	{ return dvar_type_t::DOUBLE;	}
 
 	/*************** Alias to get the value	per type	***************/
-	bool		is_float_ok()	const throw()	{ return value <= std::numeric_limits<float>::max();	}
+	bool		is_float_ok()	const throw()	{ return value <= (double)std::numeric_limits<float>::max();	}
 	float		to_float()	const throw()	{ DBG_ASSERT( is_float_ok() );  return value;		}
 
 	/*************** Comparison function	*******************************/
